About the project
Hangman is a paper-played game where one player thinks of a word or phrase, and the other tries to guess it by guessing letters within a certain number of guesses. If the second person finds the word within the guesses given by the first person, the game is won. If not, then the game is lost. In our scenario, Hangman is played between the program and the user, where the secret words that the user must find are specified in the string in the code. Find the source code of the Hangman game above.